fix: avoid dependency cycle on Fedora
Starting with 25.2, Fedora's automated test suite began failing for
Cloud images[0] due to a dependency loop. This was due to the
`Before=auditd.service` directive added to cloud-init-local.service -
auditd.service in Fedora is configured with `DefaultDependencies=no` and
`Before=sysinit.target`, but cloud-init-local.service does not include
`DefaultDependencies=no`.

This patch adds Fedora to most of the if blocks that include its
derivative distributions. The exceptions are the one waiting for
firewalld.service, since currently Fedora Cloud doesn't include it,
and the one that requires dbus.socket, as that runs after
sysinit.target.

[0] https://bugzilla.redhat.com/show_bug.cgi?id=2390898

Fixes: cc8d1b4c464d ("chore: make auditd wait for cloud-init-local.service (#6138)")
Signed-off-by: Jeremy Cline <jeremycline@linux.microsoft.com>